Extract advisor brief source mapping
